wip: rename module task into step
