+++ src/models/input.rs
@@ -1,6 +1,6 @@
 use crate::models::budget::StepBudgets;
 use crate::models::goal::Goal;
-use crate::models::task::Step;
+use crate::models::step::Step;
 use chrono::prelude::*;
 use serde::Deserialize;
 use std::collections::BTreeMap;
